# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test task API Documentation",
        "description": "Test task API Documentation Description",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Local API server"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "description": "User login api.",
                "operationId": "authLogin",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "User's phone number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "+998912223344"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "user12345"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!"
                    },
                    "422": {
                        "description": "Validation error!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User logout",
                "description": "User logout.",
                "operationId": "userLogout",
                "responses": {
                    "200": {
                        "description": "Success!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/branches": {
            "get": {
                "tags": [
                    "Branches Routes"
                ],
                "summary": "Branches list",
                "description": "Branches list.",
                "operationId": "branchList",
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Branches Routes"
                ],
                "summary": "Branch add",
                "description": "Branch add.",
                "operationId": "BranchCreate",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Branch name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Branch Name"
                    },
                    {
                        "name": "brand_id",
                        "in": "query",
                        "description": "Brand Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "district_id",
                        "in": "query",
                        "description": "District Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "12"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "422": {
                        "description": "Validation error!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/branches/{id}": {
            "get": {
                "tags": [
                    "Branches Routes"
                ],
                "summary": "Branch show",
                "description": "Branch show.",
                "operationId": "BranchShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Branch id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Branch not found!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Branches Routes"
                ],
                "summary": "Branch update",
                "description": "Branch update.",
                "operationId": "BranchUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Branch id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "143"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Branch name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "New Branch Name"
                    },
                    {
                        "name": "brand_id",
                        "in": "query",
                        "description": "New Brand Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "32"
                    },
                    {
                        "name": "district_id",
                        "in": "query",
                        "description": "New District Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "3"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Branches Routes"
                ],
                "summary": "Branch delete",
                "description": "Branch delete info.",
                "operationId": "BranchDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Branch id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "143"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Branch not found!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Delete error!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}